feat: created a route for process_prompt function:
- I have created it in a seperate folder so that our main file is as clean as possible (routes/process_prompt_route)
- retained the routes for general poem functions in routes/poem_routes
- There is a type error with get_function_definitions() which is maintained in state.py but used in process_prompt_route.py
